requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-core>=0.1.7
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
faiss-cpu==1.7.4
tiktoken==0.5.2
openai>=1.10.0
jinja2==3.1.2
aiofiles==23.2.1
pydantic>=2.5.0
